+++ firmware/prototyper/src/platform/boot.rs
@@ -15,6 +15,7 @@ use crate::sbi;
 use crate::sbi::SbiDispatcher;
 use crate::sbi::cppc::SbiCppc;
 use crate::sbi::dbtr::SbiDbtr;
+use crate::sbi::fwft::SbiFwft;
 use crate::sbi::hsm::SbiHsm;
 use crate::sbi::rfence::SbiRFence;
 use crate::sbi::sta::SbiSta;
@@ -36,6 +37,7 @@ pub fn init_board(fdt_address: usize) {
     let console = sbi::console::init(&board);
     let cppc = Some(SbiCppc::new());
     let dbtr = Some(SbiDbtr);
+    let fwft = Some(SbiFwft);
     // Get other info that later platform initialization depends on.
     let cpu_list = board.discover_misc(&tree);
     publish_cpu_enabled(cpu_list);
@@ -55,7 +57,7 @@ pub fn init_board(fdt_address: usize) {
     // so that harts observing `READY` also observe the published dispatcher.
     sbi::SBI_DISPATCHER.call_once(|| {
         SbiDispatcher::new(
-            console, cppc, dbtr, ipi, hsm, reset, rfence, susp, pmu, sta, mpxy,
+            console, cppc, dbtr, fwft, ipi, hsm, reset, rfence, susp, pmu, sta, mpxy,
         )
     });